refactor: move JiraClient to business/
