+++ src/server.ts
@@ -1,7 +1,7 @@
 import express from 'express';
 
 import conf from './ConfigurationSingleton';
-import JiraClient from './JiraClient';
+import JiraClient from './business/JiraClient';
 import authRoutesFactory from './routes/auth';
 import apiRoutesFactory from './routes/api';
 import errorHandlerMiddleware from './middlewares/errorHandler';
